:package: endpoint de SignUp em operaçao
