+++ semana12/labeddit/src/constants/endpointBase.js
@@ -1,0 +1,3 @@
+export const BASE_URL = 'https://labeddit.herokuapp.com'
+
+export const BASE_HEADERS = { 'Content-Type': 'application/json' }
